+++ fatiando/gravmag/__init__.py
@@ -29,6 +29,8 @@ The inversion modules use the forward modeling models and the
   planting anomalous densities
 * :mod:`~fatiando.gravmag.euler`: 3D Euler deconvolution methods to estimate source
   location
+* :mod:`~fatiando.gravmag.magdir`: Inversion methods to estimate the total 
+  magnetization vector of multiple sources.
 
 **Processing**
 
@@ -48,7 +50,3 @@ after modeling.
 ----
 
 """
-
-from fatiando.gravmag import (basin2d, polyprism, prism, talwani, transform,
-    harvester, sphere, tensor, fourier, imaging, euler, tesseroid,
-    half_sph_shell, eqlayer)