+++ fatiando/utils.py
@@ -6,6 +6,10 @@ Miscellaneous utility functions and classes.
 * :func:`~fatiando.utils.normal`
 * :func:`~fatiando.utils.gaussian`
 * :func:`~fatiando.utils.gaussian2d`
+* :func:`~fatiando.utils.safe_solve`
+* :func:`~fatiando.utils.safe_dot`
+* :func:`~fatiando.utils.safe_diagonal`
+* :func:`~fatiando.utils.safe_inverse`
 
 **Point scatter generation**
 
@@ -33,6 +37,7 @@ Miscellaneous utility functions and classes.
 
 **Others**
 
+* :func:`~fatiando.utils.fromimage`: Load a matrix from an image file
 * :func:`~fatiando.utils.contaminate`: Contaminate a vector with pseudo-random
   Gaussian noise
 * :func:`~fatiando.utils.dircos`: Get the 3 coordinates of a unit vector
@@ -55,8 +60,12 @@ Miscellaneous utility functions and classes.
 import math
 
 import numpy
+import scipy.sparse
+import scipy.sparse.linalg
+import scipy.misc
+import PIL.Image
 
-import fatiando.constants
+from . import constants, gridder
 
 import obspy
 
@@ -65,6 +74,143 @@ try:
     from obspy import *
 except:
     pass
+
+def fromimage(fname, ranges=None, shape=None):
+    """
+    Load an array of normalized gray-scale values from an image file.
+
+    The values will be in the range [0, 1]. The shape of the array is the shape
+    of the image (ny, nx), i.e., number of pixels in vertical (height) and
+    horizontal (width) dimensions.
+
+    Parameters:
+
+    * fname : str
+        Name of the image file
+    * ranges : [vmax, vmin] = floats
+        If not ``None``, will set the gray-scale values to this range.
+    * shape : (ny, nx)
+        If not ``None``, will interpolate the array to match this new shape
+
+    Returns:
+
+    * values : 2d-array
+        The array of gray-scale values
+
+    """
+    image = scipy.misc.fromimage(PIL.Image.open(fname), flatten=True)
+    # Invert the color scale and normalize
+    values = (image.max() - image)/numpy.abs(image).max()
+    if ranges is not None:
+        vmin, vmax = ranges
+        values *= vmax - vmin
+        values += vmin
+    if shape is not None and tuple(shape) != values.shape:
+        ny, nx = values.shape
+        X, Y = numpy.meshgrid(range(nx), range(ny))
+        values = gridder.interp(X.ravel(), Y.ravel(), values.ravel(),
+            shape)[2].reshape(shape)
+    return values
+
+def safe_inverse(matrix):
+    """
+    Calculate the inverse of a matrix using an apropriate algorithm.
+
+    Uses the standard :func:`numpy.linalg.inv` if *matrix* is dense.
+    If it is sparse (from :mod:`scipy.sparse`) then will use
+    :func:`scipy.sparse.linalg.inv`.
+
+    Parameters:
+
+    * matrix : 2d-array
+        The matrix
+
+    Returns:
+
+    * inverse : 2d-array
+        The inverse of *matrix*
+
+    """
+    if scipy.sparse.issparse(matrix):
+        return scipy.sparse.linalg.inv(matrix)
+    else:
+        return numpy.linalg.inv(matrix)
+
+def safe_solve(matrix, vector):
+    """
+    Solve a linear system using an apropriate algorithm.
+
+    Uses the standard :func:`numpy.linalg.solve` if both *matrix* and *vector*
+    are dense.
+
+    If any of the two is sparse (from :mod:`scipy.sparse`) then will use the
+    Conjugate Gradient Method (:func:`scipy.sparse.cgs`).
+
+    Parameters:
+
+    * matrix : 2d-array
+        The matrix defining the linear system
+    * vector : 1d or 2d-array
+        The right-side vector of the system
+
+    Returns:
+
+    * solution : 1d or 2d-array
+        The solution of the linear system
+
+
+    """
+    if scipy.sparse.issparse(matrix) or scipy.sparse.issparse(vector):
+        estimate, status = scipy.sparse.linalg.cgs(matrix, vector)
+        if status >= 0:
+            return estimate
+        else:
+            raise ValueError('CGS exited with input error')
+    else:
+        return numpy.linalg.solve(matrix, vector)
+
+def safe_dot(a, b):
+    """
+    Make the dot product using the appropriate method.
+
+    If *a* and *b* are dense, will use :func:`numpy.dot`. If either is sparse
+    (from :mod:`scipy.sparse`) will use the multiplication operator (i.e., \*).
+
+    Parameters:
+
+    * a, b : array or matrix
+        The vectors/matrices to take the dot product of.
+
+    Returns:
+
+    * prod : array or matrix
+        The dot product of *a* and *b*
+
+    """
+    if scipy.sparse.issparse(a) or scipy.sparse.issparse(b):
+        return a*b
+    else:
+        return numpy.dot(a, b)
+
+def safe_diagonal(matrix):
+    """
+    Get the diagonal of a matrix using the appropriate method.
+
+    Parameters:
+
+    * matrix : 2d-array, matrix, sparse matrix
+        The matrix...
+
+    Returns:
+
+    * diag : 1d-array
+        A numpy array with the diagonal of the matrix
+
+    """
+    if scipy.sparse.issparse(matrix):
+        return numpy.array(matrix.diagonal())
+    else:
+        return numpy.diagonal(matrix).copy()
 
 def vecnorm(vectors):
     """
@@ -110,7 +256,7 @@ def sph2cart(lon, lat, height):
 
     """
     d2r = numpy.pi/180.0
-    radius = fatiando.constants.MEAN_EARTH_RADIUS + height
+    radius = constants.MEAN_EARTH_RADIUS + height
     x = numpy.cos(d2r*lat)*numpy.cos(d2r*lon)*radius
     y = numpy.cos(d2r*lat)*numpy.sin(d2r*lon)*radius
     z = numpy.sin(d2r*lat)*radius
@@ -131,7 +277,7 @@ def si2nt(value):
         The value in nanoTesla
 
     """
-    return value*fatiando.constants.T2NT
+    return value*constants.T2NT
 
 def nt2si(value):
     """
@@ -148,7 +294,7 @@ def nt2si(value):
         The value in SI
 
     """
-    return value/fatiando.constants.T2NT
+    return value/constants.T2NT
 
 def si2eotvos(value):
     """
@@ -165,7 +311,7 @@ def si2eotvos(value):
         The value in Eotvos
 
     """
-    return value*fatiando.constants.SI2EOTVOS
+    return value*constants.SI2EOTVOS
 
 def eotvos2si(value):
     """
@@ -182,7 +328,7 @@ def eotvos2si(value):
         The value in SI
 
     """
-    return value/fatiando.constants.SI2EOTVOS
+    return value/constants.SI2EOTVOS
 
 def si2mgal(value):
     """
@@ -199,7 +345,7 @@ def si2mgal(value):
         The value in mGal
 
     """
-    return value*fatiando.constants.SI2MGAL
+    return value*constants.SI2MGAL
 
 def mgal2si(value):
     """
@@ -216,7 +362,7 @@ def mgal2si(value):
         The value in SI
 
     """
-    return value/fatiando.constants.SI2MGAL
+    return value/constants.SI2MGAL
 
 def vec2ang(vector):
     """
@@ -504,16 +650,16 @@ def year2sec(years):
     return 31557600.0*float(years)
 
 def contaminate(data, stddev, percent=False, return_stddev=False, seed=None):
-    """
+    r"""
     Add pseudorandom gaussian noise to an array.
 
-    Noise added is normally distributed.
+    Noise added is normally distributed with zero mean.
 
     Parameters:
 
-    * data : list or array
+    * data : array or list of arrays
         Data to contaminate
-    * stddev : float
+    * stddev : float or list of floats
         Standard deviation of the Gaussian noise that will be added to *data*
     * percent : True or False
         If ``True``, will consider *stddev* as a decimal percentage and the
@@ -531,7 +677,7 @@ def contaminate(data, stddev, percent=False, return_stddev=False, seed=None):
 
     if *return_stddev* is ``False``:
 
-    * contam : array
+    * contam : array or list of arrays
         The contaminated data array
 
     else:
@@ -540,22 +686,48 @@ def contaminate(data, stddev, percent=False, return_stddev=False, seed=None):
         The contaminated data array and the standard deviation used to
         contaminate it.
 
-    """
+    Examples:
 
-    if percent:
-        stddev = stddev*max(abs(data))
-    if stddev == 0.:
-        if return_stddev:
-            return [data, stddev]
-        else:
-            return data
+    >>> import numpy as np
+    >>> data = np.ones(5)
+    >>> noisy = contaminate(data, 0.1, seed=0)
+    >>> print noisy
+    [ 1.03137726  0.89498775  0.95284582  1.07906135  1.04172782]
+    >>> noisy, std = contaminate(data, 0.05, seed=0, percent=True,
+    ...                          return_stddev=True)
+    >>> print std
+    0.05
+    >>> print noisy
+    [ 1.01568863  0.94749387  0.97642291  1.03953067  1.02086391]
+    >>> data = [np.zeros(5), np.ones(3)]
+    >>> noisy = contaminate(data, [0.1, 0.2], seed=0)
+    >>> print noisy[0]
+    [ 0.03137726 -0.10501225 -0.04715418  0.07906135  0.04172782]
+    >>> print noisy[1]
+    [ 0.81644754  1.20192079  0.98163167]
+
+    """
     numpy.random.seed(seed)
-    noise = numpy.random.normal(scale=stddev, size=len(data))
-    # Subtract the mean so that the noise doesn't introduce a systematic shift
-    # in the data
-    noise -= noise.mean()
-    contam = numpy.array(data) + noise
+    # Check if dealing with an array or list of arrays
+    if not isinstance(stddev, list):
+        stddev = [stddev]
+        data = [data]
+    contam = []
+    for i in xrange(len(stddev)):
+        if stddev[i] == 0.:
+            contam.append(data[i])
+            continue
+        if percent:
+            stddev[i] = stddev[i]*max(abs(data[i]))
+        noise = numpy.random.normal(scale=stddev[i], size=len(data[i]))
+        # Subtract the mean so that the noise doesn't introduce a systematic
+        # shift in the data
+        noise -= noise.mean()
+        contam.append(numpy.array(data[i]) + noise)
     numpy.random.seed()
+    if len(contam) == 1:
+        contam = contam[0]
+        stddev = stddev[0]
     if return_stddev:
         return [contam, stddev]
     else: