+++ doc/changelog.rst
@@ -3,13 +3,35 @@
 Changelog
 =========
 
-Version 0.2
+Version 0.3
 -----------
 
 **Release date**:
 
 **Changes**:
 
+* **New** module :ref:`fatiando.gravmag.magdir <fatiando_gravmag_magdir>` for
+  estimating the total magnetization vector of multiple sources.
+  (`PR 87 <https://github.com/leouieda/fatiando/pull/87>`_)
+
+Version 0.2
+-----------
+
+**Release date**: 2014-01-15
+
+**Changes**:
+
+* Complete re-implementation of :ref:`fatiando.inversion <fatiando_inversion>`
+  and all modules that depended on it. Inversion routines now have a standard
+  interface. (`PR 72 <https://github.com/leouieda/fatiando/pull/72>`_)
+* Added moving window solution for Euler deconvolution in
+  :ref:`fatiando.gravmag.euler <fatiando_gravmag_euler>`.
+  (`PR 85 <https://github.com/leouieda/fatiando/pull/85>`_)
+* Renamed the ``fatiando.io`` module to
+  :ref:`fatiando.datasets <fatiando_datasets>`
+  (`PR 82 <https://github.com/leouieda/fatiando/pull/82>`_)
+* :ref:`fatiando.utils.contaminate <fatiando_utils>` can now take multiple data
+  vectors and stddevs
 * 2x speed-up of :ref:`fatiando.gravmag.talwani <fatiando_gravmag_talwani>`
   with smarter numpy array usage. (`PR 57
   <https://github.com/leouieda/fatiando/pull/57>`_)